AI-Writing-Detector/data-generation/pyproject.toml
[project]
name = "data-generation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "openai>=2.16.0",
    "python-dotenv>=1.2.1",
]
